move functions to job.go
